requirements.txt
requests>=2
lxml>=3
discord.py>=0.16
